requirements.txt
Flask==3.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
tzdata
openpyxl
qrcode[pil]
Pillow
requests
